another fairly large refactor, took care of all of the compiler warnings (most of which being size_t related (we use size_t a lot now))

+++ Useless3D/src/inc/usls/App.h
@@ -55,7 +55,7 @@ namespace usls
         void                clearScene();
         const double        time() const;
         const InputState&   getInputState() const;
-        const glm::vec2&    getScreenSize() const;
+        const glm::ivec2&   getScreenSize() const;
         void                execute();
         void                close();
